requirements.txt
kivy
pytesseract
Pillow
opencv-python
numpy
requests
pyodbc
pymysql